requirements.txt
# requirements.txt


pyramid
PyYAML
venusian
zope.interface


# EOF
